requirements.txt
fastapi==0.117.1
mcp==1.17.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
starlette==0.48.0
toml==0.10.2
uvicorn==0.37.0
fastmcp>=0.1.0
